Moving files back to root of client
