+++ client/dash.html
@@ -10,7 +10,7 @@
   <div class="push-container"></div>
 </body>
 
-<script src="/build/dash.js" type="module"></script>
-<script src="/build/push.js" type="module"></script>
+<script src="/dash.js" type="module"></script>
+<script src="/push.js" type="module"></script>
 
 </html>
