+++ client/tsconfig.json
@@ -5,7 +5,6 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "outDir": "build",
     "pretty": true,
     "skipLibCheck": true,
     "lib": [
